cf-requirements.txt
gspread
python-telegram-bot
# httpx is required by python-telegram-bot and supported by Cloudflare
httpx
